% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format.R
\name{nhi_format}
\alias{nhi_format}
\title{Check NHIs for correct format.}
\usage{
nhi_format(nhi, allow_test_cases = FALSE)
}
\arguments{
\item{nhi}{A character variable of the format \code{ZYX1234} (original)
or \code{ZYX00AX} (revised).}

\item{allow_test_cases}{If \code{TRUE}, NHIs that start with \code{Z}
(reserved for testing purposes) are treated as valid. Otherwise, treat
them as invalid (almost certainly the right choice in real-world
situations).}
}
\value{
\code{'original format'}, \code{'revised format'}, or
\code{'invalid format'}.
}
\description{
\code{nhi_format} Check whether NHIs have the basic
format of 3 valid letters (excluding \code{O} and \code{I} ), followed
by 2 digits, followed by either another two digits (the original format)
or two letters (the revised format). NHIs with \code{Z} in the first
position are for testing purposes only and so unless specified otherwise,
will be treated as invalid.
}
\examples{
# assess a test value of the original format (not a real person):
nhi_format('JBX3656')

# check a formal test case of the revised format:
nhi_format('ZZZ00AX', allow_test_cases = TRUE)

}
