% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_validity.R
\name{nhi_valid}
\alias{nhi_valid}
\title{Conduct internal validity check of an NHI.}
\usage{
nhi_valid(nhi, allow_test_cases = FALSE)
}
\arguments{
\item{nhi}{A character variable of the format \code{ZYX1234} (original)
or \code{ZYX00AX} (revised).}

\item{allow_test_cases}{If \code{TRUE}, NHIs that start with \code{Z}
(reserved for testing purposes) are treated as valid. Otherwise, regard
them as invalid (almost certainly the right choice in real-world
situations).}
}
\value{
\code{TRUE} if the format is correct and the final check digit
matches the calculated value, \code{FALSE} otherwise.
}
\description{
\code{nhi_valid} Check whether the format and final check digit of a
New Zealand NHI (National Health Index number) are correct.
}
\examples{
nhi_valid('JBX3656') # TRUE

nhi_valid('JBX3657') # FALSE (check digit should be 6)

nhi_valid('ZZZ00AX', allow_test_cases = TRUE) # TRUE

nhi_valid('ZZZ00AX') # FALSE in real-world situations

}
